+++ frontend/actions/index.js
@@ -2,7 +2,8 @@ import PipelineRequest from '@shopgate/pwa-core/classes/PipelineRequest';
 import { ERROR_HANDLE_SUPPRESS } from '@shopgate/pwa-core/constants/ErrorHandleTypes';
 import { logger } from '@shopgate/pwa-core/helpers';
 import event from '@shopgate/pwa-core/classes/Event';
-import { getCartProducts, CART_PATH } from '@shopgate/engage/cart';
+import { historyReplace } from '@shopgate/engage/core';
+import { getCartProducts, CART_PATH, fetchCart } from '@shopgate/engage/cart';
 import getCart from '@shopgate/pwa-tracking/selectors/cart';
 import { track } from '@shopgate/pwa-tracking/helpers';
 import { LoadingProvider } from '@shopgate/pwa-common/providers';
@@ -13,6 +14,10 @@ import {
   requestBoltCartToken,
 } from '../action-creators';
 import { formatTransaction } from '../helpers/formatTransaction';
+import { CHECKOUT_SUCCESS_PAGE } from '../constants';
+import config from '../config';
+
+const { showLocalCheckoutSuccessPage } = config;
 
 /**
  * @returns {Function}
@@ -57,6 +62,21 @@ export const processOrder = transaction => async (dispatch, getState) => {
 
   const products = getCartProducts(getState());
   const order = formatTransaction(transaction, products);
+
+  if (showLocalCheckoutSuccessPage) {
+    track(
+      'purchase',
+      {
+        ...order,
+        meta: { source: 'app_PWA' },
+      },
+      getState()
+    );
+    dispatch(fetchCart());
+    dispatch(historyReplace({ pathname: CHECKOUT_SUCCESS_PAGE }));
+
+    return;
+  }
 
   // checkoutSuccess triggers resetHistory, fetchCart and tracking
   event.trigger('checkoutSuccess', order);