+++ frontend/constants.js
@@ -4,3 +4,4 @@ export const ERROR_BOLT_CART_TOKEN = 'ERROR_BOLT_CART_TOKEN';
 
 export const REDUX_NAMESPACE_BOLT_CART_TOKEN = '@shopgate-project/bolt-checkout/boltCartToken';
 
+export const CHECKOUT_SUCCESS_PAGE = '/checkout_success';